refactor: use array-api-negative-index
