+++ pyproject.toml
@@ -28,6 +28,7 @@ classifiers = [
 dependencies = [
   "array-api-compat>=1.11.2",
   "array-api-extra>=0.1.1",
+  "array-api-negative-index>=1.0.0",
   "numba>=0.60.0",
   "rich>=10",
   "scipy>=1.13.1",
